relocate apps
