+++ src/module/applications/bible.ts
