+++ src/module/applications/chargen.ts
