+++ src/module/applications/npcgen.ts
